e3cc3b713 — 60 文件变更
冲突：src-tauri/crates/dao/src/migrations/mod.rs（4 处）
解决：取上游合并后的 v100 结构 + 补回我方 v200 迁移

+++ src-tauri/crates/dao/src/repo/agent_session.rs
@@ -183,7 +183,7 @@ pub async fn update_agent_session_after_query(
     id: &str,
     runtime_status: &str,
     sdk_context_json: Option<&str>,
-    tokens_delta: i32,
+    tokens_delta: i64,
     cost_delta: f64,
 ) -> Result<()> {
     let model = agent_sessions::Entity::find_by_id(id)
