+++ src-tauri/crates/harness/src/platform_adapter.rs
@@ -101,7 +101,7 @@ pub trait GatewayRequestLogRepository: Send + Sync {
         model_id: Option<&str>,
         provider_id: Option<&str>,
         status_code: i32,
-        duration_ms: i32,
+        duration_ms: i64,
         request_tokens: i64,
         response_tokens: i64,
         error_message: Option<&str>,
